Project

+++ src/main/webapp/view_job.jsp
@@ -1,10 +1,12 @@
+<%@page import="java.util.List"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8"
 	pageEncoding="UTF-8"%>
 <%@ page import="com.db.DBConnect"%>
 <%@ page import="com.dao.JobDao"%>
 <%@ page import="com.entity.jobs"%>
+<%@ page import="com.entity.User"%>
 <%@ page import="java.sql.Connection"%>
-<%@ page import="java.util.List"%>
+
 <!DOCTYPE html>
 <html>
 <head>
@@ -13,12 +15,14 @@
 <%@include file="all_components/all_css.jsp"%>
 </head>
 <body style="background-color: #f0f1f2;">
-	<%@include file="all_components/navbar.jsp"%>
+	<%@ include file="all_components/navbar.jsp"%>
 
 	<div class="container">
 		<div class="row">
 			<div class="col-md-12">
 				<h5 class="text-center text-primary">All Jobs</h5>
+
+				<%-- Show session message if available --%>
 				<%
 				String msg = (String) session.getAttribute("msg");
 				if (msg != null) {
@@ -31,6 +35,7 @@
 				}
 				%>
 
+				<%-- Load all jobs from database --%>
 				<%
 				JobDao dao = new JobDao(DBConnect.getConn());
 				List<jobs> list = dao.getAllJobs();
@@ -66,10 +71,20 @@
 							<%=j.getPdate()%></h6>
 
 						<div class="text-center">
+							<%
+							if (request.getAttribute("user") != null && "admin".equals(((User) request.getAttribute("user")).getRole())) {
+							%>
+
+							<!-- Show admin-only buttons -->
 							<a href="Edit_jobs.jsp?id=<%=j.getId()%>"
 								class="btn btn-sm bg-success text-white">Edit</a> <a
 								href="delete?id=<%=j.getId()%>"
 								class="btn btn-sm bg-danger text-white">Delete</a>
+
+							<%
+							}
+							%>
+
 						</div>
 					</div>
 				</div>
@@ -77,7 +92,6 @@
 				<%
 				}
 				%>
-
 			</div>
 		</div>
 	</div>
